Move navbar css to navbar.module.css and create Navbar folder
