+++ components/Navbar/navbar.module.css
@@ -1,0 +1,34 @@
+.navbar-light,
+.navbar-dark {
+  position: fixed;
+  min-width: 100%;
+  min-height: 50px;
+}
+
+.navbar-light {
+  background-color: rgba(252, 252, 252, .97);
+  border-bottom: 0.1px solid rgba(226, 230, 236, .99);
+}
+
+.navbar-dark {
+  background-color: rgba(13, 17, 23, .97);
+  border-bottom: 0.1px solid rgba(26, 30, 36, .99);
+}
+
+@supports ((-webkit-backdrop-filter: none) or (backdrop-filter: none)) {
+  .navbar-light,
+  .navbar-dark {
+    -webkit-backdrop-filter: blur(6px);
+    backdrop-filter: blur(6px);
+  }
+
+  .navbar-light {
+    background-color: rgba(252, 252, 252, .83);
+    border-bottom: 0.1px solid rgba(226, 230, 236, .99);
+  }
+
+  .navbar-dark {
+    background-color: rgba(13, 17, 23, .81);
+    border-bottom: 0.1px solid rgba(26, 30, 36, .99);
+  }
+}
