+++ styles/globals.css
@@ -3,31 +3,3 @@ body {
   max-width: 100%;
   padding: 0;
 }
-
-.navbar-dark {
-  background-color: rgba(13, 17, 23, .98);
-  border-bottom: 0.1px solid rgba(26, 30, 36, .99);
-}
-
-@supports ((-webkit-backdrop-filter: none) or (backdrop-filter: none)) {
-  .navbar-dark {
-    background-color: rgba(13, 17, 23, .87);
-    border-bottom: 0.1px solid rgba(26, 30, 36, .99);
-    -webkit-backdrop-filter: blur(6px);
-    backdrop-filter: blur(6px);
-  }
-}
-
-.navbar-light {
-  background-color: rgba(252, 252, 252, .98);
-  border-bottom: 0.1px solid rgba(226, 230, 236, .99);
-}
-
-@supports ((-webkit-backdrop-filter: none) or (backdrop-filter: none)) {
-  .navbar-light {
-    background-color: rgba(252, 252, 252, .87);
-    border-bottom: 0.1px solid rgba(226, 230, 236, .99);
-    -webkit-backdrop-filter: blur(6px);
-    backdrop-filter: blur(6px);
-  }
-}
